Edit provider section divided in 2 parts (location and personal information)

+++ app/Http/Controllers/DashboardController.php
@@ -10,7 +10,7 @@ use Illuminate\Support\Facades\Validator;
 //----> Models
 use App\Models\Provider;
 use App\Models\Equipment;
-
+use Illuminate\Support\Facades\Storage;
 
 class DashboardController extends Controller
 {
@@ -41,6 +41,14 @@ class DashboardController extends Controller
         $provinces = App('provinces');
 
         return view('dashboard.edit.provider', compact('provider', 'provinces'));
+    }
+
+    public function provider_location_update_page(string $id){
+        
+        $provider = Provider::find($id);
+        $provinces = App('provinces');
+
+        return view('dashboard.edit.provider_location', compact('provider', 'provinces'));
     }
 
     public function provider_create_page()
@@ -105,8 +113,92 @@ class DashboardController extends Controller
         return redirect()->route('dashboard.providers')->with('success', 'Nuevo proveedor agregado correctamente');
     }
 
+    public function provider_location_update(Request $request, $id)
+    {
+        //------> Validate provider
+        $provider = Provider::find($id);
+        if(!$provider){
+            return redirect()->route('dashboard.providers')->with('error', 'No se encontro el proveedor');
+        }
+
+        //------> Validate fields
+        $validator = Validator::make($request->all(), [
+            'province' => 'string|required',
+            'department' => 'string|nullable',
+            'city' => 'string|required',
+            'address' => 'string|nullable|max:100'
+        ]);
+        if($validator->fails()){
+            return redirect()
+                    ->route('dashboard.providers_location.update--page')
+                    ->withErrors($validator);
+        };
+
+        //-----------> Configs
+        $provinces = App::make('provinces');
+        $province_name = $provinces[$request->province];
+        $department = null;
+
+        if($request->department) $department = explode('-', $request->department);
+        $city = explode('-', $request->city);
+
+        $provider->update([
+            'province_id' => $request->province,
+            'province' => $province_name,
+            'department_id' => ($department) ? $department[0] : null,
+            'department' => ($department) ? $department[1] : null,
+            'city_id' => $city[0],
+            'city' => $city[1],
+            'address' => $request->address
+        ]);
+
+        return redirect()->route('dashboard.providers')->with('success', 'Proveedor actualizado correctamente');
+    }
+
+    public function provider_update(Request $request, $id)
+    {
+        //------> Validate provider
+        $provider = Provider::find($id);
+        if(!$provider){
+            return redirect()->route('dashboard.providers')->with('error', 'No se encontro el proveedor');
+        }
+
+        //------> Validate fields
+        $validator = Validator::make($request->all(), [
+            'name' => 'string|required|max:50',
+            'image' => 'image|nullable|max:500',
+            'web' => 'string|nullable|max:100',
+            'mail' => 'email|nullable|max:100',
+            'phone' => 'string|nullable|max:40',
+        ]);
+        if($validator->fails()){
+            return redirect()
+                    ->route('dashboard.providers.update--page', ['id' => $provider->id])
+                    ->withErrors($validator);
+        };
+        
+        //-----> Store images
+        $image_name = null ?? $provider->image;
+        $directory = 'public/images/providers';
+        if($request->hasFile('image')){
+            $image_name = time().'.'.$request->file('image')->extension();
+            $request->file('image')->storeAs($directory, $image_name);
+            Storage::delete($directory.'/'.$provider->image);
+        };
+
+        $provider->update([
+            ...$request->all(),
+            'image' => $image_name
+        ]);
+
+        return redirect()->route('dashboard.providers')->with('success', 'Proveedor actualizado correctamente');
+    }
+
     public function delete_provider($id){
         $provider = Provider::find($id);
+        if(!$provider){
+            return redirect()->route('dashboard.providers')->with('error', 'No se encontro el proveedor');
+        }
 
         $provider->delete();
